update-dependencies-and-configurations (#1)
* chore: Update .npmrc to specify npm registry

* chore: update dependencies and upgrade React version

Add new versions for node-releases and escalade packages. 
Upgrade React and related packages to version 19.0.0 to ensure 
compatibility with the latest features and improvements. 
Update other dependencies to their latest versions for better 
performance and security.

* refactor: clean up layout component and config

Remove Navbar from the Layout component to simplify the structure. 
Ensure the TanStackRouterDevtools is still conditionally loaded 
based on the environment. Adjust the Vite configuration to 
reorder plugins for better readability.